requirements.txt
# general madules
Django >3.2,<3.3
python-decouple
Pillow
djangorestframework

# third party madules
markdown
django-filter
drf-yasg[validation]
djangorestframework-simplejwt

# deployment madules

# email third party madules
django-mail-templated

# refactoring and tests
balck
flake8